Fixed all the errors to run the app properly
