+++ src/components/Header/Header.css
@@ -7,16 +7,19 @@ nav {
   width: 100%;
   background-color: slategrey;
 }
+
 .logo {
   flex: 1;
   margin: auto;
   padding: auto;
 }
+
 .imagelogo {
   border-radius: 100%;
   width: 50px;
   height: 50px;
 }
+
 .name {
   flex: 2;
 }
